Fix crash which happens when there's reordering in the beginning of a call.
The added unittest triggers this CHECK:
https://chromium.googlesource.com/external/webrtc/+/12a63b2debb29fc562dc2fa20bac5257e0345e3f/webrtc/modules/remote_bitrate_estimator/remote_estimator_proxy.cc#146

It happens because the unwrap of the sequence number fails if the unwrappers last sequence number is small, but the newly added sequence number is large (greater than last seq num + 2^15), and therefore should have been interpreted as a reordering and a backwards wrap. Since that would mean the sequence number returned from the unwrapper would be negative, it simply returns the original sequence number instead. This causes problems later where the wrap is correctly handled, and everything breaks.

The real solution should be to correctly handle wraps, but to prevent the crash this is a reasonable workaround for now.

BUG=

Review-Url: https://codereview.webrtc.org/2157843002
Cr-Commit-Position: refs/heads/master@{#13496}