Now expect the correct number of streams in EndToEndTest.GetStats.
The rtx streams were not included in the number of expected streams
but the test passed most of the time anyway due to how the checking was done.
Flake was caused when the number of registered streams jumped passed the
number of expected send streams excluding the number of rtx streams.

BUG=webrtc:6879

Review-Url: https://codereview.webrtc.org/2580343002
Cr-Commit-Position: refs/heads/master@{#15671}