Make VideoSendStreamTest.DoesNotUtilizeUlpfecForH264WithNackEnabled useful again.
This test was inadvertently made useless in https://codereview.webrtc.org/2558303002/.
Prior to that CL, at least 100 media packets -- but no ULPFEC packets -- had to
be sent before the test claimed that ULPFEC was not utilized. This CL reinstates
that behaviour, but instead of counting packets we expect the test to time out
when ULPFEC should not be utilized.

This solution is simple in code, but increases the test duration from 70 ms
to 10000 ms. Note, however, that the 70 ms duration was for a test that didn't
test anything interesting...

TESTED=Using gtest-parallel locally: 2000 test runs without flakes.
BUG=webrtc:6851

Review-Url: https://codereview.webrtc.org/2592693002
Cr-Commit-Position: refs/heads/master@{#15859}