Finalize the support for building WebRTC library for iOS with bitcode
Initial provisioning was already done in build_ios_libs.sh to support
building the WebRTC framework or static library for iOS (tvOS, watchOS)
with bitcode. Still, the actual build configuration would need to be
modified for each and every part of the build, including 3rd-party libs.
Thus, doing that more universally, at the build/config level, would be
desirable – and actually necessary to provide the intended support.

The patch for enhancing the Chromium build configs with that specific
option was landed in https://codereview.chromium.org/2631573002

NOTRY=True
BUG=webrtc:5085

Review-Url: https://codereview.webrtc.org/2633643003
Cr-Commit-Position: refs/heads/master@{#16193}