Getting rid of "benign blocking error" log spam.
For UDP sockets, instead of calling "recv" with "MSG_PEEK" to see if the socket
is alive, just check whether or not it's been closed. It seems that, at least
on some versions of Android, with some specific conditions involving loopback
sockets, calling "recv" with a buffer that's too small for a UDP packet causes
an EFAULT error.

BUG=webrtc:6715

Review-Url: https://codereview.webrtc.org/2678353006
Cr-Commit-Position: refs/heads/master@{#16522}