Ensure the RtcEventLog parser is backwards compatible
The old video send stream configs could contain multiple SSRCs and RTX SSRCs (in case of simulcast). To ensure that the RtcEventLog parser is backwards compatible, we have to return one config for every SSRC.

Also update the parsing functions for the other config types to return the config instead of passing in an output parameter.

BUG=webrtc:7731

Review-Url: https://codereview.webrtc.org/2912113002
Cr-Commit-Position: refs/heads/master@{#18343}