Add media constraint-key constants for generating offers and answers
Similar to the existing constants for the media constraint-keys intended
for media sources, add the constants for the standard keys to generate
offers and answers.

This patch also adds a few comments to RTCMediaConstraints.h, to give
a better clue on the intended usage scope of declared media-constraint
keys and values.

BUG=webrtc:7722

Review-Url: https://codereview.webrtc.org/2908013002
Cr-Commit-Position: refs/heads/master@{#18362}